+++ blog.html
@@ -18,9 +18,21 @@
             </ul>
         </nav>
         <main>
+
+            <main class="main">
+                
+                <div class="content" id="blog">
+                    <ul id="blog-list" class="blog-list"></ul>
+                </div>
+            </main>
             <h1 class="page-title">MY BLOG</h1>
             Hello World! This is my blog.
+            <div class="container" id="container"></div>
+
         </main>
+
         <footer>© 2023 Dhanvi Ganti | All Rights Reserved</footer>
+        <script src="./src/blog.js"></script>
+
     </body>
 </html>
