--- bootcamp-milestone-1/README.md
@@ -1,4 +1,0 @@
-# Dhanvi Ganti's Personal Website
-This is a personal website built using the Hack4Impact Starter Pack!
-
-Visit it here! https://drdonv.github.io
