--- bootcamp-milestone-1/blog.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>
-            My Blog
-        </title>
-        <link rel="stylesheet" type="text/css" href="styles.css" />
-    </head>
-    <body>
-        <nav class="navbar">
-            <h1 class="logo"><a href="index.html">DG</a></h1>
-            <ul class="nav-list">
-              <li class="nav"><a href="index.html">Home</a></li>
-              <li class="nav"><a href="blog.html">Blog</a></li>
-              <li class="nav"><a href="portfolio.html">Portfolio</a></li>
-              <li class="nav"><a href="resume.html">Resume</a></li>
-              <li class="nav"><a href="contact.html">Contact</a></li>
-            </ul>
-        </nav>
-        <main>
-
-            <main class="main">
-                
-                <div class="content" id="blog">
-                    <ul id="blog-list" class="blog-list"></ul>
-                </div>
-            </main>
-            <h1 class="page-title">MY BLOG</h1>
-            Hello World! This is my blog.
-            <div class="container" id="container"></div>
-
-        </main>
-
-        <footer>© 2023 Dhanvi Ganti | All Rights Reserved</footer>
-        <script src="./src/blog.js"></script>
-
-    </body>
-</html>
