--- bootcamp-milestone-1/contact.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>
-            Contact Me
-        </title>
-        <link rel="stylesheet" type="text/css" href="styles.css" />
-    </head>
-    <body>
-        <nav class="navbar">
-            <h1 class="logo"><a href="index.html">DG</a></h1>
-            <ul class="nav-list">
-                <li class="nav"><a href="index.html">Home</a></li>
-                <li class="nav"><a href="blog.html">Blog</a></li>
-                <li class="nav"><a href="portfolio.html">Portfolio</a></li>
-                <li class="nav"><a href="resume.html">Resume</a></li>
-                <li class="nav"><a href="contact.html">Contact</a></li>
-            </ul>
-        </nav>
-        <main>
-            <h1 class="page-title">CONTACT ME</h1>
-            You can contact me at dganti@calpoly.edu.
-
-            <main>
-                <p>Contact Form</p>
-                <form id="contact-form">
-                  <!--This is a comment in HTML. It lets you write text that
-                  won't be rendered in the webpage.-->
-                  <label for="name">Name</label>
-                  <input type="text" id="name" name="name" placeholder="Name" required />
-                  <br><br>
-                  <!--The "for" attribute references the "name" attribute in <input>.
-                  The "name" attribute is necessary once we start tying in JavaScript-->
-                  <label for="email">Email</label>
-                  <input type="email" id="email" name="email" placeholder="Email" required/>
-                  <br><br>
-                  </select>
-                  <label for="message">Message</label>
-                  <textarea
-                    id="message"
-                    name="message"
-                    placeholder="Message"
-                    required
-                  ></textarea>
-                  <br><br>
-                  <a href="mailto:dganti@calpoly.edu"> <input type="submit" value="Submit" /> </a>
-
-        </main>
-        <footer>© 2023 Dhanvi Ganti | All Rights Reserved</footer>
-    </body>
-</html>
