--- bootcamp-milestone-1/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>
-            Dhanvi's website
-        </title>
-        <link rel="stylesheet" type="text/css" href="styles.css" />
-    </head>
-    <body class="index">
-        <nav class="navbar">
-            <h1 class="logo"><a href="index.html">DG</a></h1>
-            <ul class="nav-list">
-                <li class="nav"><a href="index.html">Home</a></li>
-                <li class="nav"><a href="blog.html">Blog</a></li>
-                <li class="nav"><a href="portfolio.html">Portfolio</a></li>
-                <li class="nav"><a href="resume.html">Resume</a></li>
-                <li class="nav"><a href="contact.html">Contact</a></li>
-            </ul>
-        </nav>
-        <main>
-            <h1 class="page-title">ABOUT ME</h1>
-            <div class="about">
-                <div class="about-image"><img src="nature.avif" alt="Tree in a beautiful sunset"></div>
-                <div class="about-text">
-                    <p>I am <strong>Dhanvi Ganti</strong>, and I am a <strong>freshman</strong> at <strong>CalPoly San Luis Obispo</strong>.</p>
-                    <p>As a <em><strong>computer science</em></strong> major, I like using my laptop to build fun scripts and I'm into card magic.</p>
-                </div>
-            </div>
-        </main>
-        <footer>© 2023 Dhanvi Ganti | All Rights Reserved</footer>
-    </body>
-</html>
