--- bootcamp-milestone-1/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "drdonv.github.io",
-  "version": "1.0.0",
-  "description": "This is a personal website built using the Hack4Impact Starter Pack!",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "typescript": "^5.2.2"
-  }
-}
