--- bootcamp-milestone-1/portfolio.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>
-            My Work
-        </title>
-        <link rel="stylesheet" type="text/css" href="styles.css" />
-    </head>
-    <body>
-        <nav class="navbar">
-            <h1 class="logo"><a href="index.html">DG</a></h1>
-            <ul class="nav-list">
-                <li class="nav"><a href="index.html">Home</a></li>
-                <li class="nav"><a href="blog.html">Blog</a></li>
-                <li class="nav"><a href="portfolio.html">Portfolio</a></li>
-                <li class="nav"><a href="resume.html">Resume</a></li>
-                <li class="nav"><a href="contact.html">Contact</a></li>
-            </ul>
-        </nav>
-        <main>
-            <h1 class="page-title">MY WORK</h1>
-            <div class="project"><a href=index.html><img src="work.png" alt="Homepage screenshot to show in portfolio." width="500" length="500"></a></div>
-            <div class="project-details">
-                <p class="project-name"><strong>Personal Website</strong></p>
-                <p class="project-description">This is a test website made in HTML and CSS with Hack4Impact's starter pack. This project was intended for me to gain HTML/CSS experience.</p>
-                <a href="index.html">Learn More</a>
-            </div>
-            <br><br>
-            This page will be updated soon with my work experience!
-        </main>
-        <footer>© 2023 Dhanvi Ganti | All Rights Reserved</footer>
-    </body>
-</html>
