--- bootcamp-milestone-1/resume.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>
-            Resume
-        </title>
-        <link rel="stylesheet" type="text/css" href="styles.css" />
-    </head>
-    <body>
-        <nav class="navbar">
-            <h1 class="logo"><a href="index.html">DG</a></h1>
-            <ul class="nav-list">
-                <li class="nav"><a href="index.html">Home</a></li>
-                <li class="nav"><a href="blog.html">Blog</a></li>
-                <li class="nav"><a href="portfolio.html">Portfolio</a></li>
-                <li class="nav"><a href="resume.html">Resume</a></li>
-                <li class="nav"><a href="contact.html">Contact</a></li>
-            </ul>
-        </nav>
-        <main>
-            <h1 class="page-title">RESUME</h1>
-            <a href="resume.pdf" download>Download Resume</a>
-            <div class="resume">
-                <section class="section">
-                    <h2 class="section-title">Education</h2>
-                    <div class="entry">
-                        <h3 class="entry-info">California Polytechnic State University, San Luis Obispo</h3>
-                        <p class="entry-title">Computer Science Major, 2023-27 (expected)</p>
-                    </div>
-                    <div class="entry">
-                        <h3 class="entry-info">BASIS Independent Silicon Valley, San Jose</h3>
-                        <p class="entry-title">High School Diploma, 2019-2023</p>
-                    </div>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">Experience</h2>
-                    <div class="entry">
-                        <h3 class="entry-title">FRC Robotics Team 6822 Vice President</h3>
-                        <p class="entry-info">Vice President | June 2021-2022</p>
-                        <p class="entry-description">Competed in a state-wide competition against multiple other robots.<br>Learned to lead a team in terms of scheduling, deadlines, and work delegation.</p>
-                    </div>
-                    <div>
-                        <h3 class="entry-title">Accenture Learn2Lead Internship</h>
-                            <p class="entry-info">Intern</p>
-                            <p class="entry-description">Completed a 6 week survey course focused on introducing the professional culture of global business and skills required to function at a high level.<br>Developed core skills like time management, public speaking, professional communication, and engaged listening.</p>
-                    </div>
-                    <div>
-                        <h3 class="entry-title">ProcDNA</h>
-                            <p class="entry-info">Intern</p>
-                            <p class="entry-description">Using Tableau, developed business reporting solutions and product sales dashboards to track the performance of a simulated multi-billion dollar drug.</p>
-                    </div>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">Skills</h2>
-                    <ul class="skill-list">
-                        <li>Python</li>
-                        <li>Java</li>
-                        <li>JavaScript</li>
-                        <li>Swift</li>
-                        <li>Telugu</li>
-                    </ul>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">Projects</h2>
-                    <div class="entry">
-                        <h3 class="entry-title">FRC Made Easy</h3>
-                        <p class="entry-info"><a href="https://frc-rtd.readthedocs.io/en/latest/">FRC Made Easy</a></p>
-                        <p class="entry-description">A project to help complete beginners in FRC go from zero to functioning knowledge to code a robot.</p>
-                    </div>
-                    <div class="entry">
-                        <h3 class="entry-title">Flip for Philanthropy</h3>
-                        <p class="entry-info">Founder</a></p>
-                        <p class="entry-description">Coded automation scripts to help purchase high-demand goods from vendors. Then, resell acquired items for a profit, and donate the proceeds to charity.</p>
-                    </div>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">Coursework</h2>
-                    <ul class="course-list">CSC 202 - Data Structures<br>CSC 123 - Introduction to Computing</ul>
-                </section>
-                <!-- <section class="section">
-                    <h2 class="section-title">Activities</h2>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">Awards</h2>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">Interests</h2>
-                </section>
-                <section class="section">
-                    <h2 class="section-title">References</h2>
-                </section>-->
-            </div>
-            To access my resume, please contact me for a full document.
-        </main>
-        <footer>© 2023 Dhanvi Ganti | All Rights Reserved</footer>
-    </body>
-</html>
