--- bootcamp-milestone-1/styles.css
@@ -1,133 +1,0 @@
-a:link {
-  color: red;
-  text-decoration: underline;
-}
-
-a:visited {
-  color: white;
-}
-
-a:active {
-  color: red;
-}
-
-a:hover {
-  color: green;
-  text-decoration: none;
-}
-
-input:focus {
-  border: 1px solid #aaa;
-}
-.page-title {
-  font-family: 'Trebuchet MS', sans-serif;
-  margin-top: 150px;
-  margin-left: 44%;
-  
-}
-body {
-  font-size: 18px;
-  background-size: 1400px;
-  background-color: black;
-}
-
-body.index {
-  background-image: url("background.jpeg");
-}
-
-li, p {
-  font-family: monospace;
-  font-weight: bold;
-  font-size: 14px;
-  color: white;
-}
-a, p,
-h1,
-.titles {
-  color: white;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 75%;
-}
-
-h2, h3, p, ul {
-  color: white;
-}
-
-.logo {
-  padding-top: 20px;
-  padding-right: 20px;
-  padding-left: 20px;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 75%;
-}
-
-img {
-  padding: 10px;
-  border-width: 10px;
-  border-style: solid;
-  border-color: white;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 50%;
-  height: 300px;
-  width: auto;
-}
-.about-text {
-  width: 50%;
-  height: auto;
-  margin-left: auto;
-  margin-right: auto;
-}
-.logo {
-  font-family: Papyrus, fantasy;
-  display: inline-block;
-  width: 30px;
-  height: 30px;
-  margin-left: 5%
-}
-
-ul.nav-list {
-  display: inline;
-  margin-left: 10%
-
-}
-li.nav {
-  display: inline-block;
-  font-family: 'Trebuchet MS', sans-serif;
-  font-size: 24px;
-  color: white;
-  padding: 10px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: white;
-  margin-left: 100px;
-}
-footer {
-  margin-top: 50px;
-  color: white;
-}
-
-label {
-  color: white
-}
-
-ul.course-list {
-  font-family: 'Trebuchet MS', sans-serif;
-  color: lightgreen;
-}
-
-.entry-title {
-  margin-left: 40px;
-  color: lightblue;
-}
-
-.skill-list {
-  color: orange;
-  font-weight: bold;
-}
-
