--- bootcamp-milestone-2/src/app/blogData.ts
@@ -1,20 +1,0 @@
-export interface Blog {
-    
-}
-
-const blogs: Blog[] = [
-	{
-		title: "Tip 2",
-		date: "10-22-2023",
-		description: "Don't panic if you're broke - get money.",
-        slug: "blog2",
-	},
-	{
-		title: "Tip 1",
-		date: "10-21-2023",
-		description: "Don't become broke.",
-        slug: "blog1",
-	},
-];
-
-export default blogs; // Allows access to data anywhere
