+++ src/blog.ts
@@ -7,23 +7,24 @@ type Blog = {
 
 const blogs: Blog[] = [
 	{
-		title: "Tip 2",
+		title: "Unique interview skills",
 		date: "10-22-2023",
-		description: "Don't panic if you're broke - get money.",
-        slug: "blog2",
+		description: "Skills that would be really cool to show off in a a job interview.",
+        slug: "blog2.html",
 	},
 	{
-		title: "Tip 1",
+		title: "Bagels",
 		date: "10-21-2023",
-		description: "Don't become broke.",
-        slug: "blog1",
+		description: "Ranking bagel shops best to worst.",
+        slug: "blog1.html",
 	},
 ];
 
 function displayBlog() {
     const item = document.getElementById("blog")
     blogs.forEach(({title, date, description, slug}) => {
-        
+    
+        const blogList = document.getElementById("blog-list");
         const blogpost = document.createElement('div')
         const header = document.createElement('h1')
         header.innerHTML = title
@@ -35,12 +36,18 @@ function displayBlog() {
         desc.innerHTML = description
 
         const link = document.createElement('a')
-        link.href = `blog/${slug}.html`
+        link.href = `blogposts/${slug}`
+        link.innerHTML = "Learn More";
 
         blogpost.appendChild(header)
         blogpost.appendChild(date_var)
         blogpost.appendChild(desc)
         blogpost.appendChild(link)
+
+        if (blogList) {
+            blogList.appendChild(blogpost)
+        }
+
 
         item?.appendChild(blogpost)
     })
